+++ templates/tutorialv2/export/ebook/chapter.html
@@ -1,0 +1,24 @@
+{% load emarkdown %}
+
+{# this template will be used to generate the HTML file for each container ! #}
+
+{% if container.introduction %}
+    {{ container.get_introduction|emarkdown:is_js }}
+{% endif %}
+
+{% for extract in container.children %}
+    <h2 id="{{ extract.position_in_parent }}-{{ extract.slug }}">
+        <a href="#{{ extract.position_in_parent }}-{{ extract.slug }}">
+            {{ extract.title }}
+        </a>
+    </h2>
+    {% if extract.text %}
+        {{ extract.get_text|emarkdown:is_js }}
+    {% endif %}
+{% endfor %}
+
+<hr />
+
+{% if container.conclusion %}
+    {{ container.get_conclusion|emarkdown:is_js }}
+{% endif %}
